+++ app/Http/Controllers/api/TaskCommentController.php
@@ -22,19 +22,19 @@ class TaskCommentController extends Controller
     public function index()
     {
         try {
-        Log::Debug('TaskCommentController@index');
+            Log::Debug('TaskCommentController@index');
 
-        $elements = TaskComment::all(); // SELECT * FROM task_comments
+            $elements = TaskComment::all(); // SELECT * FROM task_comments
 
             return response()->json($elements, 200);
         
         } catch (\Exception $e) {
 
             Log::Error('BoardController@index', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }       
@@ -46,19 +46,23 @@ class TaskCommentController extends Controller
     public function show($id)
     {
         try {
-        Log::Debug("TaskCommentController@show $id");
+            Log::Debug("TaskCommentController@show $id");
 
-            $element = TaskComment::findOrFail($id); // SELECT * FROM task_comments WHERE id = $id 
+            $element = TaskComment::find($id); // SELECT * FROM task_comments WHERE id = $id 
 
-            return response()->json($element, 200);
+            if ($element) {
+                return response()->json($element, 200);
+            } else {
+                return response()->json(['status' => 404, 'message' => "TaskComment $id not found"], 404);
+            }
 
         } catch (\Exception $e) {
 
             Log::Error('BoardController@show', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
@@ -70,39 +74,39 @@ class TaskCommentController extends Controller
     public function store(Request $request)
     {
         try {
-        Log::Debug('TaskCommentController@store');
+            Log::Debug('TaskCommentController@store');
 
-        $validator = Validator::make($request->all(), [
-            "text" => '',
-			"from_email" => 'required|string|max:128|email',
-			"task_id" => 'required|exists:tasks,id',
+            $validator = Validator::make($request->all(), [
+                "text" => '',
+				"from_email" => 'required|string|max:128|email',
+				"task_id" => 'required|exists:tasks,id',
 
-        ]);
+            ]);
 
-        if ($validator->fails()) {
-            $data = [
-                'status' => 422,
-                'errors' => $validator->errors(),
-                'message' => 'Validation failed',
-            ];
-            Log::Debug('TaskCommentController@store validation failed', $data);
+            if ($validator->fails()) {
+                $data = [
+                    'status' => 422,
+                    'errors' => $validator->errors(),
+                    'message' => 'Validation failed',
+                ];
+                Log::Debug('TaskCommentController@store validation failed', $data);
 
-            return response()->json($data, 422);
-        }
+                return response()->json($data, 422);
+            }
 
-        $element = new TaskComment;
-        $element->text = $request->text;
-		$element->from_email = $request->from_email;
-		$element->task_id = $request->task_id;
+            $element = new TaskComment;
+            $element->text = $request->text;
+			$element->from_email = $request->from_email;
+			$element->task_id = $request->task_id;
 
             $element->save();
 
 
-        $data = [
-            'status' => 200,
+            $data = [
+                'status' => 200,
                 'task_comment' => $element
-        ];
-        Log::Debug('TaskCommentController@store saved in database', $data);
+            ];            
+            Log::Debug('TaskCommentController@store saved in database', $data);
             return response()->json($element, 200);
 
         } catch (\Exception $e) {
@@ -120,51 +124,55 @@ class TaskCommentController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, int $id)
+    public function update(Request $request, $id)
     {
         try {
-        Log::Debug("TaskCommentController@update $id");
+            Log::Debug("TaskCommentController@update $id");
 
-        $validator = Validator::make($request->all(), [
-            "text" => '',
-			"from_email" => 'string|max:128|email',
-			"task_id" => 'exists:tasks,id',
+            $validator = Validator::make($request->all(), [
+                "text" => '',
+				"from_email" => 'string|max:128|email',
+				"task_id" => 'exists:tasks,id',
 
-        ]);
+            ]);
 
-        if ($validator->fails()) {
-            $data = [
-                'status' => 422,
-                'errors' => $validator->errors(),
-                'message' => 'Validation failed',
-            ];
-            Log::Debug('TaskCommentController@store validation failed', $data);
+            if ($validator->fails()) {
+                $data = [
+                    'status' => 422,
+                    'errors' => $validator->errors(),
+                    'message' => 'Validation failed',
+                ];
+                Log::Debug('TaskCommentController@store validation failed', $data);
 
-            return response()->json($data, 422);
-        }
+                return response()->json($data, 422);
+            }
 
-            $element = TaskComment::findOrFail($id);
-        if ($request->text) {
-			$element->text = $request->text;
-		}
-		if ($request->from_email) {
-			$element->from_email = $request->from_email;
-		}
-		if ($request->task_id) {
-			$element->task_id = $request->task_id;
-		}
+            $element = TaskComment::find($id);
+            if (!$element) {
+                return response()->json(['status' => 404, 'message' => "TaskComment $id not found"], 404);
+            }
 
-        $element->save();
+            if ($request->text) {
+				$element->text = $request->text;
+			}
+			if ($request->from_email) {
+				$element->from_email = $request->from_email;
+			}
+			if ($request->task_id) {
+				$element->task_id = $request->task_id;
+			}
+
+            $element->save();
 
             return response()->json($element, 200);
 
         } catch (\Exception $e) {
 
             Log::Error('BoardController@update', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
@@ -176,9 +184,12 @@ class TaskCommentController extends Controller
     public function destroy($id)
     {
         try {
-        Log::Debug("TaskCommentController@delete $id");
+            Log::Debug("TaskCommentController@delete $id");
 
-            $element = TaskComment::findOrFail($id);
+            $element = TaskComment::find($id);
+            if (!$element) {
+                return response()->json(['status' => 404, 'message' => "TaskComment $id not found"], 404);
+            }
 
             $element->delete();
 
@@ -192,10 +203,10 @@ class TaskCommentController extends Controller
         } catch (\Exception $e) {
 
             Log::Error('BoardController@destroy', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
